+++ packages/api/server/controllers/townController/findJusticeReaders.ts
@@ -1,14 +1,15 @@
-import permissionService from '#server/services/permission';
+import shantytownService from '#server/services/shantytown';
 
 const ERROR_RESPONSES = {
     fetch_failed: { code: 400, message: 'Une lecture en base de données a échoué' },
     undefined: { code: 500, message: 'Une erreur inconnue est survenue' },
 };
 
-export default async (req, res, next) => {
-    let usersWhoCanAccessJustice;
+export default async (req, res, next): Promise<void> => {
     try {
-        usersWhoCanAccessJustice = await permissionService.findPermissionsToAccessJustice(req.params.id);
+        res.status(200).send(
+            await shantytownService.findJusticeReaders(req.params.id),
+        );
     } catch (error) {
         const { code, message } = ERROR_RESPONSES[error && error.code] || ERROR_RESPONSES.undefined;
         res.status(code).send({
@@ -16,7 +17,6 @@ export default async (req, res, next) => {
                 user_message: message,
             },
         });
-        return next(error.nativeError || error);
+        next(error.nativeError || error);
     }
-    return res.status(200).send(usersWhoCanAccessJustice);
 };
