--- packages/api/server/services/permission/index.ts
@@ -1,5 +1,0 @@
-import findPermissionsToAccessJustice from './findPermissionsToAccessJustice';
-
-export default {
-    findPermissionsToAccessJustice,
-};
