+++ packages/api/server/services/shantytown/findJusticeReaders.ts
@@ -2,13 +2,8 @@ import userModel from '#server/models/userModel';
 import ServiceError from '#server/errors/ServiceError';
 
 export default async (shantytownId: number) => {
-    let usersWhoCanAccessJustice = [];
     try {
-        usersWhoCanAccessJustice = await userModel.findJusticeReaders(shantytownId);
-        if (usersWhoCanAccessJustice.length > 0) {
-            return usersWhoCanAccessJustice;
-        }
-        return [];
+        return await userModel.findJusticeReaders(shantytownId);
     } catch (error) {
         throw new ServiceError('fetch_failed', error);
     }
