+++ packages/api/server/services/shantytown/index.ts
@@ -3,6 +3,7 @@ import list from './list';
 import find from './find';
 import findAllByActor from './findAllByActor';
 import findByNavigationLog from './findByNavigationLog';
+import findJusticeReaders from './findJusticeReaders';
 import close from './close';
 import deleteTown from './deleteTown';
 import deleteComment from './deleteComment';
@@ -17,6 +18,7 @@ export default {
     find,
     findAllByActor,
     findByNavigationLog,
+    findJusticeReaders,
     close,
     deleteTown,
     deleteComment,
